Vidazoo Adapter: save and send first request time (#11821)
* Vidazoo Adapter: save and send load time

* fix: move shared code into new library

* fix: more utils functions

* fix: moved vidazoo buildRequestData into lib

* fix: moved vidazoo interpretResponse to lib

* fix: moved vidazoo buildRequests to lib

* twistdigital lib fns
