+++ test/spec/modules/twistDigitalBidAdapter_spec.js
@@ -3,9 +3,6 @@ import {
   spec as adapter,
   createDomain,
   hashCode,
-  extractPID,
-  extractCID,
-  extractSubDomain,
   getStorageItem,
   setStorageItem,
   tryParseJSON,
@@ -18,6 +15,11 @@ import {useFakeTimers} from 'sinon';
 import {BANNER, VIDEO} from '../../../src/mediaTypes';
 import {config} from '../../../src/config';
 import {deepSetValue} from 'src/utils.js';
+import {
+  extractPID,
+  extractCID,
+  extractSubDomain
+} from '../../../libraries/vidazooUtils/bidderUtils.js';
 
 export const TEST_ID_SYSTEMS = ['britepoolid', 'criteoId', 'id5id', 'idl_env', 'lipb', 'netId', 'parrableId', 'pubcid', 'tdid', 'pubProvidedId'];
