+++ test/spec/modules/vidazooBidAdapter_spec.js
@@ -1,7 +1,11 @@
 import {expect} from 'chai';
 import {
   spec as adapter,
+  storage,
   createDomain,
+  webSessionId
+} from 'modules/vidazooBidAdapter.js';
+import {
   hashCode,
   extractPID,
   extractCID,
@@ -11,9 +15,9 @@ import {
   tryParseJSON,
   getUniqueDealId,
   getNextDealId,
-  getVidazooSessionId,
-  webSessionId
-} from 'modules/vidazooBidAdapter.js';
+  getTopWindowQueryParams,
+  getVidazooSessionId
+} from 'libraries/vidazooUtils/bidderUtils.js'
 import * as utils from 'src/utils.js';
 import {version} from 'package.json';
 import {useFakeTimers} from 'sinon';
@@ -205,15 +209,6 @@ const REQUEST = {
   }
 };
 
-function getTopWindowQueryParams() {
-  try {
-    const parsedUrl = utils.parseUrl(window.top.document.URL, {decodeSearchAsString: true});
-    return parsedUrl.search;
-  } catch (e) {
-    return '';
-  }
-}
-
 describe('VidazooBidAdapter', function () {
   describe('validtae spec', function () {
     it('exists and is a function', function () {
@@ -317,6 +312,7 @@ describe('VidazooBidAdapter', function () {
           gpid: '',
           prebidVersion: version,
           ptrace: '1000',
+          vdzhum: '1000',
           publisherId: '59ac17c192832d0011283fe3',
           url: 'https%3A%2F%2Fwww.greatsite.com',
           referrer: 'https://www.somereferrer.com',
@@ -436,6 +432,7 @@ describe('VidazooBidAdapter', function () {
           prebidVersion: version,
           schain: BID.schain,
           ptrace: '1000',
+          vdzhum: '1000',
           res: `${window.top.screen.width}x${window.top.screen.height}`,
           mediaTypes: [BANNER],
           uqs: getTopWindowQueryParams(),
@@ -526,6 +523,7 @@ describe('VidazooBidAdapter', function () {
         prebidVersion: version,
         schain: BID.schain,
         ptrace: '1000',
+        vdzhum: '1000',
         res: `${window.top.screen.width}x${window.top.screen.height}`,
         mediaTypes: [BANNER],
         uqs: getTopWindowQueryParams(),
@@ -800,14 +798,14 @@ describe('VidazooBidAdapter', function () {
       $$PREBID_GLOBAL$$.bidderSettings = {};
     });
     it('should get undefined vidazoo session id', function () {
-      const sessionId = getVidazooSessionId();
+      const sessionId = getVidazooSessionId(storage);
       expect(sessionId).to.be.empty;
     });
 
     it('should get vidazoo session id from storage', function () {
       const vidSid = '1234-5678';
       window.localStorage.setItem('vidSid', vidSid);
-      const sessionId = getVidazooSessionId();
+      const sessionId = getVidazooSessionId(storage);
       expect(sessionId).to.be.equal(vidSid);
     });
   });
@@ -826,15 +824,15 @@ describe('VidazooBidAdapter', function () {
     const key = 'myDealKey';
 
     it('should get the next deal id', function () {
-      const dealId = getNextDealId(key);
-      const nextDealId = getNextDealId(key);
+      const dealId = getNextDealId(storage, key);
+      const nextDealId = getNextDealId(storage, key);
       expect(dealId).to.be.equal(1);
       expect(nextDealId).to.be.equal(2);
     });
 
     it('should get the first deal id on expiration', function (done) {
       setTimeout(function () {
-        const dealId = getNextDealId(key, 100);
+        const dealId = getNextDealId(storage, key, 100);
         expect(dealId).to.be.equal(1);
         done();
       }, 200);
@@ -855,13 +853,13 @@ describe('VidazooBidAdapter', function () {
     const key = 'myKey';
     let uniqueDealId;
     beforeEach(() => {
-      uniqueDealId = getUniqueDealId(key, 0);
+      uniqueDealId = getUniqueDealId(storage, key, 0);
     })
 
     it('should get current unique deal id', function (done) {
       // waiting some time so `now` will become past
       setTimeout(() => {
-        const current = getUniqueDealId(key);
+        const current = getUniqueDealId(storage, key);
         expect(current).to.be.equal(uniqueDealId);
         done();
       }, 200);
@@ -869,7 +867,7 @@ describe('VidazooBidAdapter', function () {
 
     it('should get new unique deal id on expiration', function (done) {
       setTimeout(() => {
-        const current = getUniqueDealId(key, 100);
+        const current = getUniqueDealId(storage, key, 100);
         expect(current).to.not.be.equal(uniqueDealId);
         done();
       }, 200)
@@ -893,8 +891,8 @@ describe('VidazooBidAdapter', function () {
         shouldAdvanceTime: true,
         now
       });
-      setStorageItem('myKey', 2020);
-      const {value, created} = getStorageItem('myKey');
+      setStorageItem(storage, 'myKey', 2020);
+      const {value, created} = getStorageItem(storage, 'myKey');
       expect(created).to.be.equal(now);
       expect(value).to.be.equal(2020);
       expect(typeof value).to.be.equal('number');
@@ -905,7 +903,7 @@ describe('VidazooBidAdapter', function () {
     it('should get external stored value', function () {
       const value = 'superman'
       window.localStorage.setItem('myExternalKey', value);
-      const item = getStorageItem('myExternalKey');
+      const item = getStorageItem(storage, 'myExternalKey');
       expect(item).to.be.equal(value);
     });
